requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.5
deluge==1.3.15
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
hyperlink==17.3.1
idna==2.7
incremental==16.10.1
ipaddress==1.0.22
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
mock==2.0.0
mongoengine==0.16.3
numpy==1.13.3
PAM==0.4.2
Parsley==1.2
pbr==3.1.1
pep8==1.7.1
psutil==5.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pygame===1.9.1release
pygobject==3.26.1
pyliblzma==0.5.3
pymongo==3.7.2
pyOpenSSL==17.5.0
pyserial==3.4
python-libtorrent==1.1.5
python-telegram-bot==11.1.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
texttable==0.9.1
torbrowser-launcher==0.2.9
Twisted==17.9.0
txsocksx==0.0.0
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zenmap==7.60
zope.interface==4.3.2

